Memindahkan Class yang terlalu banyak
Memindahkan Class yang terlalu banyak ke Class yang baru yang nantinya extends ke Employee
